Refactor Basic acceptance tests and utils
