+++ airbyte-test-utils/build.gradle
@@ -4,6 +4,12 @@ plugins {
 
 dependencies {
     api project(':airbyte-db:db-lib')
+    implementation project(':airbyte-api')
+    implementation project(':airbyte-workers')
+
+    implementation 'io.temporal:temporal-sdk:1.8.1'
+
+
     api libs.junit.jupiter.api
 
     // Mark as compile only to avoid leaking transitively to connectors
