+++ airbyte-tests/src/acceptanceTests/java/io/airbyte/test/acceptance/BasicAcceptanceTests.java
@@ -5,6 +5,9 @@
 package io.airbyte.test.acceptance;
 
 import static io.airbyte.api.client.model.generated.ConnectionSchedule.TimeUnitEnum.MINUTES;
+import static io.airbyte.test.utils.AirbyteAcceptanceTestHelper.COLUMN_ID;
+import static io.airbyte.test.utils.AirbyteAcceptanceTestHelper.COLUMN_NAME;
+import static io.airbyte.test.utils.AirbyteAcceptanceTestHelper.STREAM_NAME;
 import static java.lang.Thread.sleep;
 import static org.junit.jupiter.api.Assertions.*;
 
@@ -13,42 +16,24 @@ import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Sets;
-import com.google.common.io.Resources;
 import io.airbyte.api.client.AirbyteApiClient;
-import io.airbyte.api.client.generated.JobsApi;
 import io.airbyte.api.client.invoker.generated.ApiClient;
 import io.airbyte.api.client.invoker.generated.ApiException;
 import io.airbyte.api.client.model.generated.*;
 import io.airbyte.commons.json.Jsons;
-import io.airbyte.commons.util.MoreProperties;
 import io.airbyte.db.Database;
-import io.airbyte.test.airbyte_test_container.AirbyteTestContainer;
-import io.airbyte.test.utils.DatabaseConnectionHelper;
+import io.airbyte.test.utils.AirbyteAcceptanceTestHelper;
 import io.airbyte.test.utils.PostgreSQLContainerHelper;
-import io.airbyte.workers.temporal.TemporalUtils;
-import io.airbyte.workers.temporal.scheduling.ConnectionManagerWorkflow;
+import io.airbyte.test.utils.SchemaTableNamePair;
 import io.airbyte.workers.temporal.scheduling.state.WorkflowState;
-import io.temporal.client.WorkflowClient;
-import io.temporal.serviceclient.WorkflowServiceStubs;
-import java.io.File;
 import java.io.IOException;
-import java.net.URI;
 import java.net.URISyntaxException;
-import java.nio.file.Path;
-import java.sql.SQLException;
 import java.time.Duration;
-import java.time.Instant;
 import java.util.*;
-import java.util.stream.Collectors;
-import org.jooq.JSONB;
-import org.jooq.Record;
-import org.jooq.Result;
-import org.jooq.SQLDialect;
 import org.junit.jupiter.api.*;
 import org.junit.jupiter.api.condition.DisabledIfEnvironmentVariable;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.testcontainers.containers.PostgreSQLContainer;
 import org.testcontainers.utility.MountableFile;
 
 /**
@@ -66,148 +51,44 @@ import org.testcontainers.utility.MountableFile;
                                matches = "true")
 public class BasicAcceptanceTests {
 
-  private static final Logger LOGGER = LoggerFactory.getLogger(AdvancedAcceptanceTests.class);
+  private static final Logger LOGGER = LoggerFactory.getLogger(BasicAcceptanceTests.class);
 
-  private static final String DOCKER_COMPOSE_FILE_NAME = "docker-compose.yaml";
-  // assume env file is one directory level up from airbyte-tests.
-  private final static File ENV_FILE = Path.of(System.getProperty("user.dir")).getParent().resolve(".env").toFile();
-
-  private static final String SOURCE_E2E_TEST_CONNECTOR_VERSION = "0.1.1";
-  private static final String DESTINATION_E2E_TEST_CONNECTOR_VERSION = "0.1.1";
-
-  private static final boolean IS_MAC = System.getProperty("os.name").startsWith("Mac");
-  private static final boolean USE_EXTERNAL_DEPLOYMENT =
-      System.getenv("USE_EXTERNAL_DEPLOYMENT") != null && System.getenv("USE_EXTERNAL_DEPLOYMENT").equalsIgnoreCase("true");
-
-  private static final String OUTPUT_NAMESPACE_PREFIX = "output_namespace_";
-  private static final String OUTPUT_NAMESPACE = OUTPUT_NAMESPACE_PREFIX + "${SOURCE_NAMESPACE}";
-  private static final String OUTPUT_STREAM_PREFIX = "output_table_";
-  private static final String TABLE_NAME = "id_and_name";
-  private static final String STREAM_NAME = TABLE_NAME;
-  private static final String COLUMN_ID = "id";
-  private static final String COLUMN_NAME = "name";
-  private static final String COLUMN_NAME_DATA = "_airbyte_data";
-  private static final String SOURCE_USERNAME = "sourceusername";
-  private static final String SOURCE_PASSWORD = "hunter2";
-
-  /**
-   * When the acceptance tests are run against a local instance of docker-compose these test
-   * containers are used.
-   */
-  private static PostgreSQLContainer sourcePsql;
-  private static PostgreSQLContainer destinationPsql;
-  private static AirbyteTestContainer airbyteTestContainer;
-  private static AirbyteApiClient apiClient;
+  private static final AirbyteAcceptanceTestHelper airbyteAcceptanceTestHelper = new AirbyteAcceptanceTestHelper(); // todo, init with api client
+  private static AirbyteApiClient apiClient; // todo let users pass this in
   private static UUID workspaceId;
 
-  private List<UUID> sourceIds;
-  private List<UUID> connectionIds;
-  private List<UUID> destinationIds;
-  private List<UUID> operationIds;
-
-  @SuppressWarnings("UnstableApiUsage")
   @BeforeAll
   public static void init() throws URISyntaxException, IOException, InterruptedException, ApiException {
-    sourcePsql = new PostgreSQLContainer("postgres:13-alpine")
-        .withUsername(SOURCE_USERNAME)
-        .withPassword(SOURCE_PASSWORD);
-    sourcePsql.start();
-
-    // by default use airbyte deployment governed by a test container.
-    if (!USE_EXTERNAL_DEPLOYMENT) {
-      LOGGER.info("Using deployment of airbyte managed by test containers.");
-      airbyteTestContainer = new AirbyteTestContainer.Builder(new File(Resources.getResource(DOCKER_COMPOSE_FILE_NAME).toURI()))
-          .setEnv(MoreProperties.envFileToProperties(ENV_FILE))
-          // override env VERSION to use dev to test current build of airbyte.
-          .setEnvVariable("VERSION", "dev")
-          // override to use test mounts.
-          .setEnvVariable("DATA_DOCKER_MOUNT", "airbyte_data_migration_test")
-          .setEnvVariable("DB_DOCKER_MOUNT", "airbyte_db_migration_test")
-          .setEnvVariable("WORKSPACE_DOCKER_MOUNT", "airbyte_workspace_migration_test")
-          .setEnvVariable("LOCAL_ROOT", "/tmp/airbyte_local_migration_test")
-          .setEnvVariable("LOCAL_DOCKER_MOUNT", "/tmp/airbyte_local_migration_test")
-          .build();
-      airbyteTestContainer.startBlocking();
-    } else {
-      LOGGER.info("Using external deployment of airbyte.");
-    }
-
     apiClient = new AirbyteApiClient(
         new ApiClient().setScheme("http")
             .setHost("localhost")
             .setPort(8001)
             .setBasePath("/api"));
-
     // work in whatever default workspace is present.
     workspaceId = apiClient.getWorkspaceApi().listWorkspaces().getWorkspaces().get(0).getWorkspaceId();
     LOGGER.info("workspaceId = " + workspaceId);
-
-    // log which connectors are being used.
-    final SourceDefinitionRead sourceDef = apiClient.getSourceDefinitionApi()
-        .getSourceDefinition(new SourceDefinitionIdRequestBody()
-            .sourceDefinitionId(UUID.fromString("decd338e-5647-4c0b-adf4-da0e75f5a750")));
-    final DestinationDefinitionRead destinationDef = apiClient.getDestinationDefinitionApi()
-        .getDestinationDefinition(new DestinationDefinitionIdRequestBody()
-            .destinationDefinitionId(UUID.fromString("25c5221d-dce2-4163-ade9-739ef790f503")));
-    LOGGER.info("pg source definition: {}", sourceDef.getDockerImageTag());
-    LOGGER.info("pg destination definition: {}", destinationDef.getDockerImageTag());
-
-    destinationPsql = new PostgreSQLContainer("postgres:13-alpine");
-    destinationPsql.start();
+    airbyteAcceptanceTestHelper.init(apiClient);
   }
 
   @AfterAll
   public static void end() {
-    sourcePsql.stop();
-    destinationPsql.stop();
-
-    if (airbyteTestContainer != null) {
-      airbyteTestContainer.stop();
-    }
+    airbyteAcceptanceTestHelper.stopDbAndContainers();
   }
 
   @BeforeEach
   public void setup() {
-    PostgreSQLContainerHelper.runSqlScript(MountableFile.forClasspathResource("postgres_init.sql"), sourcePsql);
-
-    destinationPsql = new PostgreSQLContainer("postgres:13-alpine");
-    destinationPsql.start();
-
-    sourceIds = Lists.newArrayList();
-    connectionIds = Lists.newArrayList();
-    destinationIds = Lists.newArrayList();
-    operationIds = Lists.newArrayList();
+    airbyteAcceptanceTestHelper.setup();
   }
 
   @AfterEach
   public void tearDown() {
-    try {
-      clearSourceDbData();
-
-      for (final UUID operationId : operationIds) {
-        deleteOperation(operationId);
-      }
-
-      for (final UUID connectionId : connectionIds) {
-        disableConnection(connectionId);
-      }
-
-      for (final UUID sourceId : sourceIds) {
-        deleteSource(sourceId);
-      }
-
-      for (final UUID destinationId : destinationIds) {
-        deleteDestination(destinationId);
-      }
-    } catch (Exception e) {
-      LOGGER.error("Error tearing down test fixtures:", e);
-    }
+    airbyteAcceptanceTestHelper.cleanup();
   }
 
   @Test
   @Order(-2)
   public void testGetDestinationSpec() throws ApiException {
-    final UUID destinationDefinitionId = getDestinationDefId();
+    final UUID destinationDefinitionId = airbyteAcceptanceTestHelper.getDestinationDefId();
     final DestinationDefinitionSpecificationRead spec = apiClient.getDestinationDefinitionSpecificationApi()
         .getDestinationDefinitionSpecification(
             new DestinationDefinitionIdWithWorkspaceId().destinationDefinitionId(destinationDefinitionId).workspaceId(UUID.randomUUID()));
@@ -227,7 +108,7 @@ public class BasicAcceptanceTests {
   @Test
   @Order(0)
   public void testGetSourceSpec() throws ApiException {
-    final UUID sourceDefId = getPostgresSourceDefinitionId();
+    final UUID sourceDefId = airbyteAcceptanceTestHelper.getPostgresSourceDefinitionId();
     final SourceDefinitionSpecificationRead spec = apiClient.getSourceDefinitionSpecificationApi()
         .getSourceDefinitionSpecification(new SourceDefinitionIdWithWorkspaceId().sourceDefinitionId(sourceDefId).workspaceId(UUID.randomUUID()));
     assertEquals(sourceDefId, spec.getSourceDefinitionId());
@@ -237,11 +118,11 @@ public class BasicAcceptanceTests {
   @Test
   @Order(1)
   public void testCreateDestination() throws ApiException {
-    final UUID destinationDefId = getDestinationDefId();
-    final JsonNode destinationConfig = getDestinationDbConfig();
+    final UUID destinationDefId = airbyteAcceptanceTestHelper.getDestinationDefId();
+    final JsonNode destinationConfig = airbyteAcceptanceTestHelper.getDestinationDbConfig();
     final String name = "AccTestDestinationDb-" + UUID.randomUUID();
 
-    final DestinationRead createdDestination = createDestination(
+    final DestinationRead createdDestination = airbyteAcceptanceTestHelper.createDestination(
         name,
         workspaceId,
         destinationDefId,
@@ -250,13 +131,13 @@ public class BasicAcceptanceTests {
     assertEquals(name, createdDestination.getName());
     assertEquals(destinationDefId, createdDestination.getDestinationDefinitionId());
     assertEquals(workspaceId, createdDestination.getWorkspaceId());
-    assertEquals(getDestinationDbConfigWithHiddenPassword(), createdDestination.getConnectionConfiguration());
+    assertEquals(airbyteAcceptanceTestHelper.getDestinationDbConfigWithHiddenPassword(), createdDestination.getConnectionConfiguration());
   }
 
   @Test
   @Order(2)
   public void testDestinationCheckConnection() throws ApiException {
-    final UUID destinationId = createDestination().getDestinationId();
+    final UUID destinationId = airbyteAcceptanceTestHelper.createDestination().getDestinationId();
 
     final CheckConnectionRead.StatusEnum checkOperationStatus = apiClient.getDestinationApi()
         .checkConnectionToDestination(new DestinationIdRequestBody().destinationId(destinationId))
@@ -269,10 +150,10 @@ public class BasicAcceptanceTests {
   @Order(3)
   public void testCreateSource() throws ApiException {
     final String dbName = "acc-test-db";
-    final UUID postgresSourceDefinitionId = getPostgresSourceDefinitionId();
-    final JsonNode sourceDbConfig = getSourceDbConfig();
+    final UUID postgresSourceDefinitionId = airbyteAcceptanceTestHelper.getPostgresSourceDefinitionId();
+    final JsonNode sourceDbConfig = airbyteAcceptanceTestHelper.getSourceDbConfig();
 
-    final SourceRead response = createSource(
+    final SourceRead response = airbyteAcceptanceTestHelper.createSource(
         dbName,
         workspaceId,
         postgresSourceDefinitionId,
@@ -290,7 +171,7 @@ public class BasicAcceptanceTests {
   @Test
   @Order(4)
   public void testSourceCheckConnection() throws ApiException {
-    final UUID sourceId = createPostgresSource().getSourceId();
+    final UUID sourceId = airbyteAcceptanceTestHelper.createPostgresSource().getSourceId();
 
     final CheckConnectionRead checkConnectionRead = apiClient.getSourceApi().checkConnectionToSource(new SourceIdRequestBody().sourceId(sourceId));
 
@@ -303,9 +184,9 @@ public class BasicAcceptanceTests {
   @Test
   @Order(5)
   public void testDiscoverSourceSchema() throws ApiException {
-    final UUID sourceId = createPostgresSource().getSourceId();
+    final UUID sourceId = airbyteAcceptanceTestHelper.createPostgresSource().getSourceId();
 
-    final AirbyteCatalog actual = discoverSourceSchema(sourceId);
+    final AirbyteCatalog actual = airbyteAcceptanceTestHelper.discoverSourceSchema(sourceId);
 
     final Map<String, Map<String, DataType>> fields = ImmutableMap.of(
         COLUMN_ID, ImmutableMap.of("type", DataType.NUMBER),
@@ -340,16 +221,17 @@ public class BasicAcceptanceTests {
   @Test
   @Order(6)
   public void testCreateConnection() throws ApiException {
-    final UUID sourceId = createPostgresSource().getSourceId();
-    final AirbyteCatalog catalog = discoverSourceSchema(sourceId);
-    final UUID destinationId = createDestination().getDestinationId();
-    final UUID operationId = createOperation().getOperationId();
+    final UUID sourceId = airbyteAcceptanceTestHelper.createPostgresSource().getSourceId();
+    final AirbyteCatalog catalog = airbyteAcceptanceTestHelper.discoverSourceSchema(sourceId);
+    final UUID destinationId = airbyteAcceptanceTestHelper.createDestination().getDestinationId();
+    final UUID operationId = airbyteAcceptanceTestHelper.createOperation().getOperationId();
     final String name = "test-connection-" + UUID.randomUUID();
     final ConnectionSchedule schedule = new ConnectionSchedule().timeUnit(MINUTES).units(100L);
     final SyncMode syncMode = SyncMode.FULL_REFRESH;
     final DestinationSyncMode destinationSyncMode = DestinationSyncMode.OVERWRITE;
     catalog.getStreams().forEach(s -> s.getConfig().syncMode(syncMode).destinationSyncMode(destinationSyncMode));
-    final ConnectionRead createdConnection = createConnection(name, sourceId, destinationId, List.of(operationId), catalog, schedule);
+    final ConnectionRead createdConnection =
+        airbyteAcceptanceTestHelper.createConnection(name, sourceId, destinationId, List.of(operationId), catalog, schedule);
 
     assertEquals(sourceId, createdConnection.getSourceId());
     assertEquals(destinationId, createdConnection.getDestinationId());
@@ -363,9 +245,9 @@ public class BasicAcceptanceTests {
   @Test
   @Order(7)
   public void testCancelSync() throws Exception {
-    final SourceDefinitionRead sourceDefinition = createE2eSourceDefinition();
+    final SourceDefinitionRead sourceDefinition = airbyteAcceptanceTestHelper.createE2eSourceDefinition();
 
-    final SourceRead source = createSource(
+    final SourceRead source = airbyteAcceptanceTestHelper.createSource(
         "E2E Test Source -" + UUID.randomUUID(),
         workspaceId,
         sourceDefinition.getSourceDefinitionId(),
@@ -377,18 +259,19 @@ public class BasicAcceptanceTests {
 
     final String connectionName = "test-connection";
     final UUID sourceId = source.getSourceId();
-    final UUID destinationId = createDestination().getDestinationId();
-    final UUID operationId = createOperation().getOperationId();
-    final AirbyteCatalog catalog = discoverSourceSchema(sourceId);
+    final UUID destinationId = airbyteAcceptanceTestHelper.createDestination().getDestinationId();
+    final UUID operationId = airbyteAcceptanceTestHelper.createOperation().getOperationId();
+    final AirbyteCatalog catalog = airbyteAcceptanceTestHelper.discoverSourceSchema(sourceId);
     final SyncMode syncMode = SyncMode.FULL_REFRESH;
     final DestinationSyncMode destinationSyncMode = DestinationSyncMode.OVERWRITE;
     catalog.getStreams().forEach(s -> s.getConfig().syncMode(syncMode).destinationSyncMode(destinationSyncMode));
     final UUID connectionId =
-        createConnection(connectionName, sourceId, destinationId, List.of(operationId), catalog, null).getConnectionId();
+        airbyteAcceptanceTestHelper.createConnection(connectionName, sourceId, destinationId, List.of(operationId), catalog, null).getConnectionId();
     final JobInfoRead connectionSyncRead = apiClient.getConnectionApi().syncConnection(new ConnectionIdRequestBody().connectionId(connectionId));
 
     // wait to get out of PENDING
-    final JobRead jobRead = waitWhileJobHasStatus(apiClient.getJobsApi(), connectionSyncRead.getJob(), Set.of(JobStatus.PENDING));
+    final JobRead jobRead =
+        AirbyteAcceptanceTestHelper.waitWhileJobHasStatus(apiClient.getJobsApi(), connectionSyncRead.getJob(), Set.of(JobStatus.PENDING));
     assertEquals(JobStatus.RUNNING, jobRead.getStatus());
 
     final var resp = apiClient.getJobsApi().cancelJob(new JobIdRequestBody().id(connectionSyncRead.getJob().getId()));
@@ -399,17 +282,18 @@ public class BasicAcceptanceTests {
   @Order(8)
   public void testScheduledSync() throws Exception {
     final String connectionName = "test-connection";
-    final UUID sourceId = createPostgresSource().getSourceId();
-    final UUID destinationId = createDestination().getDestinationId();
-    final UUID operationId = createOperation().getOperationId();
-    final AirbyteCatalog catalog = discoverSourceSchema(sourceId);
+    final UUID sourceId = airbyteAcceptanceTestHelper.createPostgresSource().getSourceId();
+    final UUID destinationId = airbyteAcceptanceTestHelper.createDestination().getDestinationId();
+    final UUID operationId = airbyteAcceptanceTestHelper.createOperation().getOperationId();
+    final AirbyteCatalog catalog = airbyteAcceptanceTestHelper.discoverSourceSchema(sourceId);
 
     final ConnectionSchedule connectionSchedule = new ConnectionSchedule().units(1L).timeUnit(MINUTES);
     final SyncMode syncMode = SyncMode.FULL_REFRESH;
     final DestinationSyncMode destinationSyncMode = DestinationSyncMode.OVERWRITE;
     catalog.getStreams().forEach(s -> s.getConfig().syncMode(syncMode).destinationSyncMode(destinationSyncMode));
 
-    var conn = createConnection(connectionName, sourceId, destinationId, List.of(operationId), catalog, connectionSchedule);
+    final var conn =
+        airbyteAcceptanceTestHelper.createConnection(connectionName, sourceId, destinationId, List.of(operationId), catalog, connectionSchedule);
 
     // When a new connection is created, Airbyte might sync it immediately (before the sync interval).
     // Then it will wait the sync interval.
@@ -417,69 +301,71 @@ public class BasicAcceptanceTests {
     // syncs in progress
     List<io.airbyte.api.client.model.generated.JobWithAttemptsRead> jobs = new ArrayList<>();
     while (jobs.size() < 2) {
-      var listSyncJobsRequest = new io.airbyte.api.client.model.generated.JobListRequestBody().configTypes(List.of(JobConfigType.SYNC))
+      final var listSyncJobsRequest = new io.airbyte.api.client.model.generated.JobListRequestBody().configTypes(List.of(JobConfigType.SYNC))
           .configId(conn.getConnectionId().toString());
-      var resp = apiClient.getJobsApi().listJobsFor(listSyncJobsRequest);
+      final var resp = apiClient.getJobsApi().listJobsFor(listSyncJobsRequest);
       jobs = resp.getJobs();
       sleep(Duration.ofSeconds(30).toMillis());
     }
 
-    assertSourceAndDestinationDbInSync(false);
+    airbyteAcceptanceTestHelper.assertSourceAndDestinationDbInSync(false);
   }
 
   @Test
   @Order(9)
   public void testMultipleSchemasAndTablesSync() throws Exception {
     // create tables in another schema
-    PostgreSQLContainerHelper.runSqlScript(MountableFile.forClasspathResource("postgres_second_schema_multiple_tables.sql"), sourcePsql);
+    PostgreSQLContainerHelper.runSqlScript(MountableFile.forClasspathResource("postgres_second_schema_multiple_tables.sql"),
+        airbyteAcceptanceTestHelper.getSourcePsql());
 
     final String connectionName = "test-connection";
-    final UUID sourceId = createPostgresSource().getSourceId();
-    final UUID destinationId = createDestination().getDestinationId();
-    final UUID operationId = createOperation().getOperationId();
-    final AirbyteCatalog catalog = discoverSourceSchema(sourceId);
+    final UUID sourceId = airbyteAcceptanceTestHelper.createPostgresSource().getSourceId();
+    final UUID destinationId = airbyteAcceptanceTestHelper.createDestination().getDestinationId();
+    final UUID operationId = airbyteAcceptanceTestHelper.createOperation().getOperationId();
+    final AirbyteCatalog catalog = airbyteAcceptanceTestHelper.discoverSourceSchema(sourceId);
 
     final SyncMode syncMode = SyncMode.FULL_REFRESH;
     final DestinationSyncMode destinationSyncMode = DestinationSyncMode.OVERWRITE;
     catalog.getStreams().forEach(s -> s.getConfig().syncMode(syncMode).destinationSyncMode(destinationSyncMode));
     final UUID connectionId =
-        createConnection(connectionName, sourceId, destinationId, List.of(operationId), catalog, null).getConnectionId();
+        airbyteAcceptanceTestHelper.createConnection(connectionName, sourceId, destinationId, List.of(operationId), catalog, null).getConnectionId();
     final JobInfoRead connectionSyncRead = apiClient.getConnectionApi().syncConnection(new ConnectionIdRequestBody().connectionId(connectionId));
-    waitForSuccessfulJob(apiClient.getJobsApi(), connectionSyncRead.getJob());
-    assertSourceAndDestinationDbInSync(false);
+    AirbyteAcceptanceTestHelper.waitForSuccessfulJob(apiClient.getJobsApi(), connectionSyncRead.getJob());
+    airbyteAcceptanceTestHelper.assertSourceAndDestinationDbInSync(false);
   }
 
   @Test
   @Order(10)
   public void testMultipleSchemasSameTablesSync() throws Exception {
     // create tables in another schema
-    PostgreSQLContainerHelper.runSqlScript(MountableFile.forClasspathResource("postgres_separate_schema_same_table.sql"), sourcePsql);
+    PostgreSQLContainerHelper.runSqlScript(MountableFile.forClasspathResource("postgres_separate_schema_same_table.sql"),
+        airbyteAcceptanceTestHelper.getSourcePsql());
 
     final String connectionName = "test-connection";
-    final UUID sourceId = createPostgresSource().getSourceId();
-    final UUID destinationId = createDestination().getDestinationId();
-    final UUID operationId = createOperation().getOperationId();
-    final AirbyteCatalog catalog = discoverSourceSchema(sourceId);
+    final UUID sourceId = airbyteAcceptanceTestHelper.createPostgresSource().getSourceId();
+    final UUID destinationId = airbyteAcceptanceTestHelper.createDestination().getDestinationId();
+    final UUID operationId = airbyteAcceptanceTestHelper.createOperation().getOperationId();
+    final AirbyteCatalog catalog = airbyteAcceptanceTestHelper.discoverSourceSchema(sourceId);
 
     final SyncMode syncMode = SyncMode.FULL_REFRESH;
     final DestinationSyncMode destinationSyncMode = DestinationSyncMode.OVERWRITE;
     catalog.getStreams().forEach(s -> s.getConfig().syncMode(syncMode).destinationSyncMode(destinationSyncMode));
     final UUID connectionId =
-        createConnection(connectionName, sourceId, destinationId, List.of(operationId), catalog, null).getConnectionId();
+        airbyteAcceptanceTestHelper.createConnection(connectionName, sourceId, destinationId, List.of(operationId), catalog, null).getConnectionId();
 
     final JobInfoRead connectionSyncRead = apiClient.getConnectionApi().syncConnection(new ConnectionIdRequestBody().connectionId(connectionId));
-    waitForSuccessfulJob(apiClient.getJobsApi(), connectionSyncRead.getJob());
-    assertSourceAndDestinationDbInSync(false);
+    AirbyteAcceptanceTestHelper.waitForSuccessfulJob(apiClient.getJobsApi(), connectionSyncRead.getJob());
+    airbyteAcceptanceTestHelper.assertSourceAndDestinationDbInSync(false);
   }
 
   @Test
   @Order(11)
   public void testIncrementalDedupeSync() throws Exception {
     final String connectionName = "test-connection";
-    final UUID sourceId = createPostgresSource().getSourceId();
-    final UUID destinationId = createDestination().getDestinationId();
-    final UUID operationId = createOperation().getOperationId();
-    final AirbyteCatalog catalog = discoverSourceSchema(sourceId);
+    final UUID sourceId = airbyteAcceptanceTestHelper.createPostgresSource().getSourceId();
+    final UUID destinationId = airbyteAcceptanceTestHelper.createDestination().getDestinationId();
+    final UUID operationId = airbyteAcceptanceTestHelper.createOperation().getOperationId();
+    final AirbyteCatalog catalog = airbyteAcceptanceTestHelper.discoverSourceSchema(sourceId);
     final SyncMode syncMode = SyncMode.INCREMENTAL;
     final DestinationSyncMode destinationSyncMode = DestinationSyncMode.APPEND_DEDUP;
     catalog.getStreams().forEach(s -> s.getConfig()
@@ -488,32 +374,32 @@ public class BasicAcceptanceTests {
         .destinationSyncMode(destinationSyncMode)
         .primaryKey(List.of(List.of(COLUMN_NAME))));
     final UUID connectionId =
-        createConnection(connectionName, sourceId, destinationId, List.of(operationId), catalog, null).getConnectionId();
+        airbyteAcceptanceTestHelper.createConnection(connectionName, sourceId, destinationId, List.of(operationId), catalog, null).getConnectionId();
 
     // sync from start
     final JobInfoRead connectionSyncRead1 = apiClient.getConnectionApi()
         .syncConnection(new ConnectionIdRequestBody().connectionId(connectionId));
-    waitForSuccessfulJob(apiClient.getJobsApi(), connectionSyncRead1.getJob());
+    AirbyteAcceptanceTestHelper.waitForSuccessfulJob(apiClient.getJobsApi(), connectionSyncRead1.getJob());
 
-    assertSourceAndDestinationDbInSync(true);
+    airbyteAcceptanceTestHelper.assertSourceAndDestinationDbInSync(true);
 
     // add new records and run again.
-    final Database source = getSourceDatabase();
-    final List<JsonNode> expectedRawRecords = retrieveSourceRecords(source, STREAM_NAME);
+    final Database source = airbyteAcceptanceTestHelper.getSourceDatabase();
+    final List<JsonNode> expectedRawRecords = airbyteAcceptanceTestHelper.retrieveSourceRecords(source, STREAM_NAME);
     expectedRawRecords.add(Jsons.jsonNode(ImmutableMap.builder().put(COLUMN_ID, 6).put(COLUMN_NAME, "sherif").build()));
     expectedRawRecords.add(Jsons.jsonNode(ImmutableMap.builder().put(COLUMN_ID, 7).put(COLUMN_NAME, "chris").build()));
     source.query(ctx -> ctx.execute("UPDATE id_and_name SET id=6 WHERE name='sherif'"));
     source.query(ctx -> ctx.execute("INSERT INTO id_and_name(id, name) VALUES(7, 'chris')"));
     // retrieve latest snapshot of source records after modifications; the deduplicated table in
     // destination should mirror this latest state of records
-    final List<JsonNode> expectedNormalizedRecords = retrieveSourceRecords(source, STREAM_NAME);
+    final List<JsonNode> expectedNormalizedRecords = airbyteAcceptanceTestHelper.retrieveSourceRecords(source, STREAM_NAME);
 
     final JobInfoRead connectionSyncRead2 = apiClient.getConnectionApi()
         .syncConnection(new ConnectionIdRequestBody().connectionId(connectionId));
-    waitForSuccessfulJob(apiClient.getJobsApi(), connectionSyncRead2.getJob());
+    AirbyteAcceptanceTestHelper.waitForSuccessfulJob(apiClient.getJobsApi(), connectionSyncRead2.getJob());
 
-    assertRawDestinationContains(expectedRawRecords, new SchemaTableNamePair("public", STREAM_NAME));
-    assertNormalizedDestinationContains(expectedNormalizedRecords);
+    airbyteAcceptanceTestHelper.assertRawDestinationContains(expectedRawRecords, new SchemaTableNamePair("public", STREAM_NAME));
+    airbyteAcceptanceTestHelper.assertNormalizedDestinationContains(expectedNormalizedRecords);
   }
 
   @Test
@@ -521,10 +407,10 @@ public class BasicAcceptanceTests {
   public void testIncrementalSync() throws Exception {
     LOGGER.info("Starting testIncrementalSync()");
     final String connectionName = "test-connection";
-    final UUID sourceId = createPostgresSource().getSourceId();
-    final UUID destinationId = createDestination().getDestinationId();
-    final UUID operationId = createOperation().getOperationId();
-    final AirbyteCatalog catalog = discoverSourceSchema(sourceId);
+    final UUID sourceId = airbyteAcceptanceTestHelper.createPostgresSource().getSourceId();
+    final UUID destinationId = airbyteAcceptanceTestHelper.createDestination().getDestinationId();
+    final UUID operationId = airbyteAcceptanceTestHelper.createOperation().getOperationId();
+    final AirbyteCatalog catalog = airbyteAcceptanceTestHelper.discoverSourceSchema(sourceId);
     final AirbyteStream stream = catalog.getStreams().get(0).getStream();
 
     assertEquals(Lists.newArrayList(SyncMode.FULL_REFRESH, SyncMode.INCREMENTAL), stream.getSupportedSyncModes());
@@ -540,20 +426,20 @@ public class BasicAcceptanceTests {
         .cursorField(List.of(COLUMN_ID))
         .destinationSyncMode(destinationSyncMode));
     final UUID connectionId =
-        createConnection(connectionName, sourceId, destinationId, List.of(operationId), catalog, null).getConnectionId();
+        airbyteAcceptanceTestHelper.createConnection(connectionName, sourceId, destinationId, List.of(operationId), catalog, null).getConnectionId();
     LOGGER.info("Beginning testIncrementalSync() sync 1");
 
     final JobInfoRead connectionSyncRead1 = apiClient.getConnectionApi()
         .syncConnection(new ConnectionIdRequestBody().connectionId(connectionId));
-    waitForSuccessfulJob(apiClient.getJobsApi(), connectionSyncRead1.getJob());
+    AirbyteAcceptanceTestHelper.waitForSuccessfulJob(apiClient.getJobsApi(), connectionSyncRead1.getJob());
     LOGGER.info("state after sync 1: {}", apiClient.getConnectionApi().getState(new ConnectionIdRequestBody().connectionId(connectionId)));
 
-    assertSourceAndDestinationDbInSync(false);
+    airbyteAcceptanceTestHelper.assertSourceAndDestinationDbInSync(false);
 
     // add new records and run again.
-    final Database source = getSourceDatabase();
+    final Database source = airbyteAcceptanceTestHelper.getSourceDatabase();
     // get contents of source before mutating records.
-    final List<JsonNode> expectedRecords = retrieveSourceRecords(source, STREAM_NAME);
+    final List<JsonNode> expectedRecords = airbyteAcceptanceTestHelper.retrieveSourceRecords(source, STREAM_NAME);
     expectedRecords.add(Jsons.jsonNode(ImmutableMap.builder().put(COLUMN_ID, 6).put(COLUMN_NAME, "geralt").build()));
     // add a new record
     source.query(ctx -> ctx.execute("INSERT INTO id_and_name(id, name) VALUES(6, 'geralt')"));
@@ -565,31 +451,31 @@ public class BasicAcceptanceTests {
     LOGGER.info("Starting testIncrementalSync() sync 2");
     final JobInfoRead connectionSyncRead2 = apiClient.getConnectionApi()
         .syncConnection(new ConnectionIdRequestBody().connectionId(connectionId));
-    waitForSuccessfulJob(apiClient.getJobsApi(), connectionSyncRead2.getJob());
+    AirbyteAcceptanceTestHelper.waitForSuccessfulJob(apiClient.getJobsApi(), connectionSyncRead2.getJob());
     LOGGER.info("state after sync 2: {}", apiClient.getConnectionApi().getState(new ConnectionIdRequestBody().connectionId(connectionId)));
 
-    assertRawDestinationContains(expectedRecords, new SchemaTableNamePair("public", STREAM_NAME));
+    airbyteAcceptanceTestHelper.assertRawDestinationContains(expectedRecords, new SchemaTableNamePair("public", STREAM_NAME));
 
     // reset back to no data.
 
     LOGGER.info("Starting testIncrementalSync() reset");
     final JobInfoRead jobInfoRead = apiClient.getConnectionApi().resetConnection(new ConnectionIdRequestBody().connectionId(connectionId));
-    waitWhileJobHasStatus(apiClient.getJobsApi(), jobInfoRead.getJob(),
+    AirbyteAcceptanceTestHelper.waitWhileJobHasStatus(apiClient.getJobsApi(), jobInfoRead.getJob(),
         Sets.newHashSet(JobStatus.PENDING, JobStatus.RUNNING, JobStatus.INCOMPLETE, JobStatus.FAILED));
 
     LOGGER.info("state after reset: {}", apiClient.getConnectionApi().getState(new ConnectionIdRequestBody().connectionId(connectionId)));
 
-    assertRawDestinationContains(Collections.emptyList(), new SchemaTableNamePair("public",
+    airbyteAcceptanceTestHelper.assertRawDestinationContains(Collections.emptyList(), new SchemaTableNamePair("public",
         STREAM_NAME));
 
     // sync one more time. verify it is the equivalent of a full refresh.
     LOGGER.info("Starting testIncrementalSync() sync 3");
     final JobInfoRead connectionSyncRead3 =
         apiClient.getConnectionApi().syncConnection(new ConnectionIdRequestBody().connectionId(connectionId));
-    waitForSuccessfulJob(apiClient.getJobsApi(), connectionSyncRead3.getJob());
+    AirbyteAcceptanceTestHelper.waitForSuccessfulJob(apiClient.getJobsApi(), connectionSyncRead3.getJob());
     LOGGER.info("state after sync 3: {}", apiClient.getConnectionApi().getState(new ConnectionIdRequestBody().connectionId(connectionId)));
 
-    assertSourceAndDestinationDbInSync(false);
+    airbyteAcceptanceTestHelper.assertSourceAndDestinationDbInSync(false);
 
   }
 
@@ -597,10 +483,10 @@ public class BasicAcceptanceTests {
   @Order(13)
   public void testDeleteConnection() throws Exception {
     final String connectionName = "test-connection";
-    final UUID sourceId = createPostgresSource().getSourceId();
-    final UUID destinationId = createDestination().getDestinationId();
-    final UUID operationId = createOperation().getOperationId();
-    final AirbyteCatalog catalog = discoverSourceSchema(sourceId);
+    final UUID sourceId = airbyteAcceptanceTestHelper.createPostgresSource().getSourceId();
+    final UUID destinationId = airbyteAcceptanceTestHelper.createDestination().getDestinationId();
+    final UUID operationId = airbyteAcceptanceTestHelper.createOperation().getOperationId();
+    final AirbyteCatalog catalog = airbyteAcceptanceTestHelper.discoverSourceSchema(sourceId);
     final SyncMode syncMode = SyncMode.INCREMENTAL;
     final DestinationSyncMode destinationSyncMode = DestinationSyncMode.APPEND_DEDUP;
     catalog.getStreams().forEach(s -> s.getConfig()
@@ -610,17 +496,18 @@ public class BasicAcceptanceTests {
         .primaryKey(List.of(List.of(COLUMN_NAME))));
 
     UUID connectionId =
-        createConnection(connectionName, sourceId, destinationId, List.of(operationId), catalog, null).getConnectionId();
+        airbyteAcceptanceTestHelper.createConnection(connectionName, sourceId, destinationId, List.of(operationId), catalog, null).getConnectionId();
 
     final JobInfoRead connectionSyncRead = apiClient.getConnectionApi().syncConnection(new ConnectionIdRequestBody().connectionId(connectionId));
-    waitWhileJobHasStatus(apiClient.getJobsApi(), connectionSyncRead.getJob(), Set.of(JobStatus.RUNNING));
+    AirbyteAcceptanceTestHelper.waitWhileJobHasStatus(apiClient.getJobsApi(), connectionSyncRead.getJob(), Set.of(JobStatus.RUNNING));
 
     // test normal deletion of connection
     LOGGER.info("Calling delete connection...");
     apiClient.getConnectionApi().deleteConnection(new ConnectionIdRequestBody().connectionId(connectionId));
 
     // remove connection to avoid exception during tear down
-    connectionIds.remove(connectionId);
+    // connectionIds.remove(connectionId); // todo remove
+    airbyteAcceptanceTestHelper.removeConnection(connectionId);
 
     LOGGER.info("Waiting for connection to be deleted...");
     Thread.sleep(500);
@@ -635,9 +522,10 @@ public class BasicAcceptanceTests {
 
     // test deletion of connection when temporal workflow is in a bad state
     LOGGER.info("Testing connection deletion when temporal is in a terminal state");
-    connectionId = createConnection(connectionName, sourceId, destinationId, List.of(operationId), catalog, null).getConnectionId();
+    connectionId =
+        airbyteAcceptanceTestHelper.createConnection(connectionName, sourceId, destinationId, List.of(operationId), catalog, null).getConnectionId();
 
-    terminateTemporalWorkflow(connectionId);
+    airbyteAcceptanceTestHelper.terminateTemporalWorkflow(connectionId);
 
     // we should still be able to delete the connection when the temporal workflow is in this state
     apiClient.getConnectionApi().deleteConnection(new ConnectionIdRequestBody().connectionId(connectionId));
@@ -660,10 +548,10 @@ public class BasicAcceptanceTests {
     // test just ensures that the underlying workflow
     // is running after the update method is called.
     final String connectionName = "test-connection";
-    final UUID sourceId = createPostgresSource().getSourceId();
-    final UUID destinationId = createDestination().getDestinationId();
-    final UUID operationId = createOperation().getOperationId();
-    final AirbyteCatalog catalog = discoverSourceSchema(sourceId);
+    final UUID sourceId = airbyteAcceptanceTestHelper.createPostgresSource().getSourceId();
+    final UUID destinationId = airbyteAcceptanceTestHelper.createDestination().getDestinationId();
+    final UUID operationId = airbyteAcceptanceTestHelper.createOperation().getOperationId();
+    final AirbyteCatalog catalog = airbyteAcceptanceTestHelper.discoverSourceSchema(sourceId);
     final SyncMode syncMode = SyncMode.INCREMENTAL;
     final DestinationSyncMode destinationSyncMode = DestinationSyncMode.APPEND_DEDUP;
     catalog.getStreams().forEach(s -> s.getConfig()
@@ -673,17 +561,19 @@ public class BasicAcceptanceTests {
         .primaryKey(List.of(List.of(COLUMN_NAME))));
 
     LOGGER.info("Testing connection update when temporal is in a terminal state");
-    final UUID connectionId = createConnection(connectionName, sourceId, destinationId, List.of(operationId), catalog, null).getConnectionId();
+    final UUID connectionId =
+        airbyteAcceptanceTestHelper.createConnection(connectionName, sourceId, destinationId, List.of(operationId), catalog, null).getConnectionId();
 
-    terminateTemporalWorkflow(connectionId);
+    airbyteAcceptanceTestHelper.terminateTemporalWorkflow(connectionId);
 
     // we should still be able to update the connection when the temporal workflow is in this state
-    updateConnectionSchedule(connectionId, new ConnectionSchedule().timeUnit(ConnectionSchedule.TimeUnitEnum.HOURS).units(1L));
+    airbyteAcceptanceTestHelper.updateConnectionSchedule(connectionId,
+        new ConnectionSchedule().timeUnit(ConnectionSchedule.TimeUnitEnum.HOURS).units(1L));
 
     LOGGER.info("Waiting for workflow to be recreated...");
     Thread.sleep(500);
 
-    final WorkflowState workflowState = getWorkflowState(connectionId);
+    final WorkflowState workflowState = airbyteAcceptanceTestHelper.getWorkflowState(connectionId);
     assertTrue(workflowState.isRunning());
   }
 
@@ -693,8 +583,8 @@ public class BasicAcceptanceTests {
     // This test only covers the specific behavior of updating a connection that does not have an
     // underlying temporal workflow.
     final String connectionName = "test-connection";
-    final SourceDefinitionRead sourceDefinition = createE2eSourceDefinition();
-    final SourceRead source = createSource(
+    final SourceDefinitionRead sourceDefinition = airbyteAcceptanceTestHelper.createE2eSourceDefinition();
+    final SourceRead source = airbyteAcceptanceTestHelper.createSource(
         "E2E Test Source -" + UUID.randomUUID(),
         workspaceId,
         sourceDefinition.getSourceDefinitionId(),
@@ -704,9 +594,9 @@ public class BasicAcceptanceTests {
             .put("message_interval", 100)
             .build()));
     final UUID sourceId = source.getSourceId();
-    final UUID destinationId = createDestination().getDestinationId();
-    final UUID operationId = createOperation().getOperationId();
-    final AirbyteCatalog catalog = discoverSourceSchema(sourceId);
+    final UUID destinationId = airbyteAcceptanceTestHelper.createDestination().getDestinationId();
+    final UUID operationId = airbyteAcceptanceTestHelper.createOperation().getOperationId();
+    final AirbyteCatalog catalog = airbyteAcceptanceTestHelper.discoverSourceSchema(sourceId);
     final SyncMode syncMode = SyncMode.INCREMENTAL;
     final DestinationSyncMode destinationSyncMode = DestinationSyncMode.APPEND_DEDUP;
     catalog.getStreams().forEach(s -> s.getConfig()
@@ -716,12 +606,13 @@ public class BasicAcceptanceTests {
         .primaryKey(List.of(List.of(COLUMN_NAME))));
 
     LOGGER.info("Testing manual sync when temporal is in a terminal state");
-    final UUID connectionId = createConnection(connectionName, sourceId, destinationId, List.of(operationId), catalog, null).getConnectionId();
+    final UUID connectionId =
+        airbyteAcceptanceTestHelper.createConnection(connectionName, sourceId, destinationId, List.of(operationId), catalog, null).getConnectionId();
 
     LOGGER.info("Starting first manual sync");
     final JobInfoRead firstJobInfo = apiClient.getConnectionApi().syncConnection(new ConnectionIdRequestBody().connectionId(connectionId));
     LOGGER.info("Terminating workflow during first sync");
-    terminateTemporalWorkflow(connectionId);
+    airbyteAcceptanceTestHelper.terminateTemporalWorkflow(connectionId);
 
     LOGGER.info("Submitted another manual sync");
     apiClient.getConnectionApi().syncConnection(new ConnectionIdRequestBody().connectionId(connectionId));
@@ -729,7 +620,7 @@ public class BasicAcceptanceTests {
     LOGGER.info("Waiting for workflow to be recreated...");
     Thread.sleep(500);
 
-    final WorkflowState workflowState = getWorkflowState(connectionId);
+    final WorkflowState workflowState = airbyteAcceptanceTestHelper.getWorkflowState(connectionId);
     assertTrue(workflowState.isRunning());
     assertTrue(workflowState.isSkipScheduling());
 
@@ -744,10 +635,10 @@ public class BasicAcceptanceTests {
     // This test only covers the specific behavior of updating a connection that does not have an
     // underlying temporal workflow.
     final String connectionName = "test-connection";
-    final UUID sourceId = createPostgresSource().getSourceId();
-    final UUID destinationId = createDestination().getDestinationId();
-    final UUID operationId = createOperation().getOperationId();
-    final AirbyteCatalog catalog = discoverSourceSchema(sourceId);
+    final UUID sourceId = airbyteAcceptanceTestHelper.createPostgresSource().getSourceId();
+    final UUID destinationId = airbyteAcceptanceTestHelper.createDestination().getDestinationId();
+    final UUID operationId = airbyteAcceptanceTestHelper.createOperation().getOperationId();
+    final AirbyteCatalog catalog = airbyteAcceptanceTestHelper.discoverSourceSchema(sourceId);
     final SyncMode syncMode = SyncMode.INCREMENTAL;
     final DestinationSyncMode destinationSyncMode = DestinationSyncMode.APPEND_DEDUP;
     catalog.getStreams().forEach(s -> s.getConfig()
@@ -757,16 +648,17 @@ public class BasicAcceptanceTests {
         .primaryKey(List.of(List.of(COLUMN_NAME))));
 
     LOGGER.info("Testing reset connection when temporal is in a terminal state");
-    final UUID connectionId = createConnection(connectionName, sourceId, destinationId, List.of(operationId), catalog, null).getConnectionId();
+    final UUID connectionId =
+        airbyteAcceptanceTestHelper.createConnection(connectionName, sourceId, destinationId, List.of(operationId), catalog, null).getConnectionId();
 
-    terminateTemporalWorkflow(connectionId);
+    airbyteAcceptanceTestHelper.terminateTemporalWorkflow(connectionId);
 
     apiClient.getConnectionApi().resetConnection(new ConnectionIdRequestBody().connectionId(connectionId));
 
     LOGGER.info("Waiting for workflow to be recreated...");
     Thread.sleep(500);
 
-    final WorkflowState workflowState = getWorkflowState(connectionId);
+    final WorkflowState workflowState = airbyteAcceptanceTestHelper.getWorkflowState(connectionId);
     assertTrue(workflowState.isRunning());
     assertTrue(workflowState.isResetConnection());
   }
@@ -776,10 +668,10 @@ public class BasicAcceptanceTests {
   // See relevant issue: https://github.com/airbytehq/airbyte/issues/8397
   @Disabled
   public void testFailureTimeout() throws Exception {
-    final SourceDefinitionRead sourceDefinition = createE2eSourceDefinition();
-    final DestinationDefinitionRead destinationDefinition = createE2eDestinationDefinition();
+    final SourceDefinitionRead sourceDefinition = airbyteAcceptanceTestHelper.createE2eSourceDefinition();
+    final DestinationDefinitionRead destinationDefinition = airbyteAcceptanceTestHelper.createE2eDestinationDefinition();
 
-    final SourceRead source = createSource(
+    final SourceRead source = airbyteAcceptanceTestHelper.createSource(
         "E2E Test Source -" + UUID.randomUUID(),
         workspaceId,
         sourceDefinition.getSourceDefinitionId(),
@@ -791,7 +683,7 @@ public class BasicAcceptanceTests {
 
     // Destination fails after processing 5 messages, so the job should fail after the graceful close
     // timeout of 1 minute
-    final DestinationRead destination = createDestination(
+    final DestinationRead destination = airbyteAcceptanceTestHelper.createDestination(
         "E2E Test Destination -" + UUID.randomUUID(),
         workspaceId,
         destinationDefinition.getDestinationDefinitionId(),
@@ -803,20 +695,21 @@ public class BasicAcceptanceTests {
     final String connectionName = "test-connection";
     final UUID sourceId = source.getSourceId();
     final UUID destinationId = destination.getDestinationId();
-    final AirbyteCatalog catalog = discoverSourceSchema(sourceId);
+    final AirbyteCatalog catalog = airbyteAcceptanceTestHelper.discoverSourceSchema(sourceId);
 
     final UUID connectionId =
-        createConnection(connectionName, sourceId, destinationId, Collections.emptyList(), catalog, null)
+        airbyteAcceptanceTestHelper.createConnection(connectionName, sourceId, destinationId, Collections.emptyList(), catalog, null)
             .getConnectionId();
 
     final JobInfoRead connectionSyncRead1 = apiClient.getConnectionApi()
         .syncConnection(new ConnectionIdRequestBody().connectionId(connectionId));
 
     // wait to get out of pending.
-    final JobRead runningJob = waitWhileJobHasStatus(apiClient.getJobsApi(), connectionSyncRead1.getJob(), Sets.newHashSet(JobStatus.PENDING));
+    final JobRead runningJob =
+        AirbyteAcceptanceTestHelper.waitWhileJobHasStatus(apiClient.getJobsApi(), connectionSyncRead1.getJob(), Sets.newHashSet(JobStatus.PENDING));
 
     // wait for job for max of 3 minutes, by which time the job attempt should have failed
-    waitWhileJobHasStatus(apiClient.getJobsApi(), runningJob, Sets.newHashSet(JobStatus.RUNNING), Duration.ofMinutes(3));
+    AirbyteAcceptanceTestHelper.waitWhileJobHasStatus(apiClient.getJobsApi(), runningJob, Sets.newHashSet(JobStatus.RUNNING), Duration.ofMinutes(3));
 
     final JobIdRequestBody jobId = new JobIdRequestBody().id(runningJob.getId());
     final JobInfoRead jobInfo = apiClient.getJobsApi().getJobInfo(jobId);
@@ -828,418 +721,6 @@ public class BasicAcceptanceTests {
     } finally {
       apiClient.getJobsApi().cancelJob(jobId);
     }
-  }
-
-  private WorkflowClient getWorkflowClient() {
-    final WorkflowServiceStubs temporalService = TemporalUtils.createTemporalService("localhost:7233");
-    return WorkflowClient.newInstance(temporalService);
-  }
-
-  private WorkflowState getWorkflowState(final UUID connectionId) {
-    final WorkflowClient workflowCLient = getWorkflowClient();
-
-    // check if temporal workflow is reachable
-    final ConnectionManagerWorkflow connectionManagerWorkflow =
-        workflowCLient.newWorkflowStub(ConnectionManagerWorkflow.class, "connection_manager_" + connectionId);
-
-    return connectionManagerWorkflow.getState();
-  }
-
-  private void terminateTemporalWorkflow(final UUID connectionId) {
-    final WorkflowClient workflowCLient = getWorkflowClient();
-
-    // check if temporal workflow is reachable
-    getWorkflowState(connectionId);
-
-    // Terminate workflow
-    LOGGER.info("Terminating temporal workflow...");
-    workflowCLient.newUntypedWorkflowStub("connection_manager_" + connectionId).terminate("");
-
-    // remove connection to avoid exception during tear down
-    connectionIds.remove(connectionId);
-  }
-
-  private AirbyteCatalog discoverSourceSchema(final UUID sourceId) throws ApiException {
-    return apiClient.getSourceApi().discoverSchemaForSource(new SourceDiscoverSchemaRequestBody().sourceId(sourceId)).getCatalog();
-  }
-
-  private void assertSourceAndDestinationDbInSync(final boolean withScdTable) throws Exception {
-    final Database source = getSourceDatabase();
-
-    final Set<SchemaTableNamePair> sourceTables = listAllTables(source);
-    final Set<SchemaTableNamePair> sourceTablesWithRawTablesAdded = addAirbyteGeneratedTables(withScdTable, sourceTables);
-    final Database destination = getDestinationDatabase();
-    final Set<SchemaTableNamePair> destinationTables = listAllTables(destination);
-    assertEquals(sourceTablesWithRawTablesAdded, destinationTables,
-        String.format("streams did not match.\n source stream names: %s\n destination stream names: %s\n", sourceTables, destinationTables));
-
-    for (final SchemaTableNamePair pair : sourceTables) {
-      final List<JsonNode> sourceRecords = retrieveSourceRecords(source, pair.getFullyQualifiedTableName());
-      assertRawDestinationContains(sourceRecords, pair);
-    }
-  }
-
-  private Database getSourceDatabase() {
-    return getDatabase(sourcePsql);
-  }
-
-  private Database getDestinationDatabase() {
-    return getDatabase(destinationPsql);
-  }
-
-  private Database getDatabase(final PostgreSQLContainer db) {
-    return new Database(DatabaseConnectionHelper.createDslContext(db, SQLDialect.POSTGRES));
-  }
-
-  private Set<SchemaTableNamePair> listAllTables(final Database database) throws SQLException {
-    return database.query(
-        context -> {
-          final Result<Record> fetch =
-              context.fetch(
-                  "SELECT tablename, schemaname FROM pg_catalog.pg_tables WHERE schemaname != 'pg_catalog' AND schemaname != 'information_schema'");
-          return fetch.stream()
-              .map(record -> {
-                final var schemaName = (String) record.get("schemaname");
-                final var tableName = (String) record.get("tablename");
-                return new SchemaTableNamePair(schemaName, tableName);
-              })
-              .collect(Collectors.toSet());
-        });
-  }
-
-  private Set<SchemaTableNamePair> addAirbyteGeneratedTables(final boolean withScdTable, final Set<SchemaTableNamePair> sourceTables) {
-    return sourceTables.stream().flatMap(x -> {
-      final String cleanedNameStream = x.tableName.replace(".", "_");
-      final List<SchemaTableNamePair> explodedStreamNames = new ArrayList<>(List.of(
-          new SchemaTableNamePair(OUTPUT_NAMESPACE_PREFIX + x.schemaName,
-              String.format("_airbyte_raw_%s%s", OUTPUT_STREAM_PREFIX, cleanedNameStream)),
-          new SchemaTableNamePair(OUTPUT_NAMESPACE_PREFIX + x.schemaName, String.format("%s%s", OUTPUT_STREAM_PREFIX, cleanedNameStream))));
-      if (withScdTable) {
-        explodedStreamNames
-            .add(new SchemaTableNamePair("_airbyte_" + OUTPUT_NAMESPACE_PREFIX + x.schemaName,
-                String.format("%s%s_stg", OUTPUT_STREAM_PREFIX, cleanedNameStream)));
-        explodedStreamNames
-            .add(new SchemaTableNamePair(OUTPUT_NAMESPACE_PREFIX + x.schemaName, String.format("%s%s_scd", OUTPUT_STREAM_PREFIX, cleanedNameStream)));
-      }
-      return explodedStreamNames.stream();
-    }).collect(Collectors.toSet());
-  }
-
-  private void assertRawDestinationContains(final List<JsonNode> sourceRecords, final SchemaTableNamePair pair) throws Exception {
-    final Set<JsonNode> destinationRecords = new HashSet<>(retrieveRawDestinationRecords(pair));
-
-    assertEquals(sourceRecords.size(), destinationRecords.size(),
-        String.format("destination contains: %s record. source contains: %s, \nsource records %s \ndestination records: %s",
-            destinationRecords.size(), sourceRecords.size(), sourceRecords, destinationRecords));
-
-    for (final JsonNode sourceStreamRecord : sourceRecords) {
-      assertTrue(destinationRecords.contains(sourceStreamRecord),
-          String.format("destination does not contain record:\n %s \n destination contains:\n %s\n",
-              sourceStreamRecord, destinationRecords));
-    }
-  }
-
-  private void assertNormalizedDestinationContains(final List<JsonNode> sourceRecords) throws Exception {
-    final Database destination = getDestinationDatabase();
-    final String finalDestinationTable = String.format("%spublic.%s%s", OUTPUT_NAMESPACE_PREFIX, OUTPUT_STREAM_PREFIX, STREAM_NAME.replace(".", "_"));
-    final List<JsonNode> destinationRecords = retrieveSourceRecords(destination, finalDestinationTable);
-
-    assertEquals(sourceRecords.size(), destinationRecords.size(),
-        String.format("destination contains: %s record. source contains: %s", sourceRecords.size(), destinationRecords.size()));
-
-    for (final JsonNode sourceStreamRecord : sourceRecords) {
-      assertTrue(
-          destinationRecords.stream()
-              .anyMatch(r -> r.get(COLUMN_NAME).asText().equals(sourceStreamRecord.get(COLUMN_NAME).asText())
-                  && r.get(COLUMN_ID).asInt() == sourceStreamRecord.get(COLUMN_ID).asInt()),
-          String.format("destination does not contain record:\n %s \n destination contains:\n %s\n", sourceStreamRecord, destinationRecords));
-    }
-  }
-
-  private ConnectionRead createConnection(final String name,
-                                          final UUID sourceId,
-                                          final UUID destinationId,
-                                          final List<UUID> operationIds,
-                                          final AirbyteCatalog catalog,
-                                          final ConnectionSchedule schedule)
-      throws ApiException {
-    final ConnectionRead connection = apiClient.getConnectionApi().createConnection(
-        new ConnectionCreate()
-            .status(ConnectionStatus.ACTIVE)
-            .sourceId(sourceId)
-            .destinationId(destinationId)
-            .syncCatalog(catalog)
-            .schedule(schedule)
-            .operationIds(operationIds)
-            .name(name)
-            .namespaceDefinition(NamespaceDefinitionType.CUSTOMFORMAT)
-            .namespaceFormat(OUTPUT_NAMESPACE)
-            .prefix(OUTPUT_STREAM_PREFIX));
-    connectionIds.add(connection.getConnectionId());
-    return connection;
-  }
-
-  private ConnectionRead updateConnectionSchedule(final UUID connectionId, final ConnectionSchedule newSchedule) throws ApiException {
-    final ConnectionRead connectionRead = apiClient.getConnectionApi().getConnection(new ConnectionIdRequestBody().connectionId(connectionId));
-
-    return apiClient.getConnectionApi().updateConnection(
-        new ConnectionUpdate()
-            .namespaceDefinition(connectionRead.getNamespaceDefinition())
-            .namespaceFormat(connectionRead.getNamespaceFormat())
-            .prefix(connectionRead.getPrefix())
-            .connectionId(connectionId)
-            .operationIds(connectionRead.getOperationIds())
-            .status(connectionRead.getStatus())
-            .syncCatalog(connectionRead.getSyncCatalog())
-            .name(connectionRead.getName())
-            .resourceRequirements(connectionRead.getResourceRequirements())
-            .schedule(newSchedule) // only field being updated
-    );
-  }
-
-  private DestinationRead createDestination() throws ApiException {
-    return createDestination(
-        "AccTestDestination-" + UUID.randomUUID(),
-        workspaceId,
-        getDestinationDefId(),
-        getDestinationDbConfig());
-  }
-
-  private DestinationRead createDestination(final String name, final UUID workspaceId, final UUID destinationDefId, final JsonNode destinationConfig)
-      throws ApiException {
-    final DestinationRead destination =
-        apiClient.getDestinationApi().createDestination(new DestinationCreate()
-            .name(name)
-            .connectionConfiguration(Jsons.jsonNode(destinationConfig))
-            .workspaceId(workspaceId)
-            .destinationDefinitionId(destinationDefId));
-    destinationIds.add(destination.getDestinationId());
-    return destination;
-  }
-
-  private OperationRead createOperation() throws ApiException {
-    final OperatorConfiguration normalizationConfig = new OperatorConfiguration()
-        .operatorType(OperatorType.NORMALIZATION).normalization(new OperatorNormalization().option(
-            OperatorNormalization.OptionEnum.BASIC));
-
-    final OperationCreate operationCreate = new OperationCreate()
-        .workspaceId(workspaceId)
-        .name("AccTestDestination-" + UUID.randomUUID()).operatorConfiguration(normalizationConfig);
-
-    final OperationRead operation = apiClient.getOperationApi().createOperation(operationCreate);
-    operationIds.add(operation.getOperationId());
-    return operation;
-  }
-
-  private UUID getDestinationDefId() throws ApiException {
-    return apiClient.getDestinationDefinitionApi().listDestinationDefinitions().getDestinationDefinitions()
-        .stream()
-        .filter(dr -> dr.getName().toLowerCase().contains("postgres"))
-        .findFirst()
-        .orElseThrow()
-        .getDestinationDefinitionId();
-  }
-
-  private List<JsonNode> retrieveSourceRecords(final Database database, final String table) throws SQLException {
-    return database.query(context -> context.fetch(String.format("SELECT * FROM %s;", table)))
-        .stream()
-        .map(Record::intoMap)
-        .map(Jsons::jsonNode)
-        .collect(Collectors.toList());
-  }
-
-  private List<JsonNode> retrieveDestinationRecords(final Database database, final String table) throws SQLException {
-    return database.query(context -> context.fetch(String.format("SELECT * FROM %s;", table)))
-        .stream()
-        .map(Record::intoMap)
-        .map(r -> r.get(COLUMN_NAME_DATA))
-        .map(f -> (JSONB) f)
-        .map(JSONB::data)
-        .map(Jsons::deserialize)
-        .map(Jsons::jsonNode)
-        .collect(Collectors.toList());
-  }
-
-  private List<JsonNode> retrieveRawDestinationRecords(final SchemaTableNamePair pair) throws Exception {
-    final Database destination = getDestinationDatabase();
-    final Set<SchemaTableNamePair> namePairs = listAllTables(destination);
-
-    final String rawStreamName = String.format("_airbyte_raw_%s%s", OUTPUT_STREAM_PREFIX, pair.tableName.replace(".", "_"));
-    final SchemaTableNamePair rawTablePair = new SchemaTableNamePair(OUTPUT_NAMESPACE_PREFIX + pair.schemaName, rawStreamName);
-    assertTrue(namePairs.contains(rawTablePair), "can't find a non-normalized version (raw) of " + rawTablePair.getFullyQualifiedTableName());
-
-    return retrieveDestinationRecords(destination, rawTablePair.getFullyQualifiedTableName());
-  }
-
-  private JsonNode getSourceDbConfig() {
-    return getDbConfig(sourcePsql, false, false);
-  }
-
-  private JsonNode getDestinationDbConfig() {
-    return getDbConfig(destinationPsql, false, true);
-  }
-
-  private JsonNode getDestinationDbConfigWithHiddenPassword() {
-    return getDbConfig(destinationPsql, true, true);
-  }
-
-  private JsonNode getDbConfig(final PostgreSQLContainer psql, final boolean hiddenPassword, final boolean withSchema) {
-    try {
-      final Map<Object, Object> dbConfig = localConfig(psql, hiddenPassword, withSchema);
-      return Jsons.jsonNode(dbConfig);
-    } catch (final Exception e) {
-      throw new RuntimeException(e);
-    }
-  }
-
-  private Map<Object, Object> localConfig(final PostgreSQLContainer psql, final boolean hiddenPassword, final boolean withSchema) {
-    final Map<Object, Object> dbConfig = new HashMap<>();
-    // don't use psql.getHost() directly since the ip we need differs depending on environment
-    if (IS_MAC) {
-      dbConfig.put("host", "host.docker.internal");
-    } else {
-      dbConfig.put("host", "localhost");
-    }
-
-    if (hiddenPassword) {
-      dbConfig.put("password", "**********");
-    } else {
-      dbConfig.put("password", psql.getPassword());
-    }
-
-    dbConfig.put("port", psql.getFirstMappedPort());
-    dbConfig.put("database", psql.getDatabaseName());
-    dbConfig.put("username", psql.getUsername());
-    dbConfig.put("ssl", false);
-
-    if (withSchema) {
-      dbConfig.put("schema", "public");
-    }
-    return dbConfig;
-  }
-
-  private SourceDefinitionRead createE2eSourceDefinition() throws ApiException {
-    return apiClient.getSourceDefinitionApi().createSourceDefinition(new SourceDefinitionCreate()
-        .name("E2E Test Source")
-        .dockerRepository("airbyte/source-e2e-test")
-        .dockerImageTag(SOURCE_E2E_TEST_CONNECTOR_VERSION)
-        .documentationUrl(URI.create("https://example.com")));
-  }
-
-  private DestinationDefinitionRead createE2eDestinationDefinition() throws ApiException {
-    return apiClient.getDestinationDefinitionApi().createDestinationDefinition(new DestinationDefinitionCreate()
-        .name("E2E Test Destination")
-        .dockerRepository("airbyte/destination-e2e-test")
-        .dockerImageTag(DESTINATION_E2E_TEST_CONNECTOR_VERSION)
-        .documentationUrl(URI.create("https://example.com")));
-  }
-
-  private SourceRead createPostgresSource() throws ApiException {
-    return createSource(
-        "acceptanceTestDb-" + UUID.randomUUID(),
-        workspaceId,
-        getPostgresSourceDefinitionId(),
-        getSourceDbConfig());
-  }
-
-  private SourceRead createSource(final String name, final UUID workspaceId, final UUID sourceDefId, final JsonNode sourceConfig)
-      throws ApiException {
-    final SourceRead source = apiClient.getSourceApi().createSource(new SourceCreate()
-        .name(name)
-        .sourceDefinitionId(sourceDefId)
-        .workspaceId(workspaceId)
-        .connectionConfiguration(sourceConfig));
-    sourceIds.add(source.getSourceId());
-    return source;
-  }
-
-  private UUID getPostgresSourceDefinitionId() throws ApiException {
-    return apiClient.getSourceDefinitionApi().listSourceDefinitions().getSourceDefinitions()
-        .stream()
-        .filter(sourceRead -> sourceRead.getName().equalsIgnoreCase("postgres"))
-        .findFirst()
-        .orElseThrow()
-        .getSourceDefinitionId();
-  }
-
-  private void clearSourceDbData() throws SQLException {
-    final Database database = getSourceDatabase();
-    final Set<SchemaTableNamePair> pairs = listAllTables(database);
-    for (final SchemaTableNamePair pair : pairs) {
-      database.query(context -> context.execute(String.format("DROP TABLE %s.%s", pair.schemaName, pair.tableName)));
-    }
-  }
-
-  private void deleteSource(final UUID sourceId) throws ApiException {
-    apiClient.getSourceApi().deleteSource(new SourceIdRequestBody().sourceId(sourceId));
-  }
-
-  private void disableConnection(final UUID connectionId) throws ApiException {
-    final ConnectionRead connection = apiClient.getConnectionApi().getConnection(new ConnectionIdRequestBody().connectionId(connectionId));
-    final ConnectionUpdate connectionUpdate =
-        new ConnectionUpdate()
-            .prefix(connection.getPrefix())
-            .connectionId(connectionId)
-            .operationIds(connection.getOperationIds())
-            .status(ConnectionStatus.DEPRECATED)
-            .schedule(connection.getSchedule())
-            .syncCatalog(connection.getSyncCatalog());
-    apiClient.getConnectionApi().updateConnection(connectionUpdate);
-  }
-
-  private void deleteDestination(final UUID destinationId) throws ApiException {
-    apiClient.getDestinationApi().deleteDestination(new DestinationIdRequestBody().destinationId(destinationId));
-  }
-
-  private void deleteOperation(final UUID destinationId) throws ApiException {
-    apiClient.getOperationApi().deleteOperation(new OperationIdRequestBody().operationId(destinationId));
-  }
-
-  private static void waitForSuccessfulJob(final JobsApi jobsApi, final JobRead originalJob) throws InterruptedException, ApiException {
-    final JobRead job = waitWhileJobHasStatus(jobsApi, originalJob, Sets.newHashSet(JobStatus.PENDING, JobStatus.RUNNING));
-
-    if (!JobStatus.SUCCEEDED.equals(job.getStatus())) {
-      // If a job failed during testing, show us why.
-      final JobIdRequestBody id = new JobIdRequestBody();
-      id.setId(originalJob.getId());
-      for (final AttemptInfoRead attemptInfo : jobsApi.getJobInfo(id).getAttempts()) {
-        LOGGER.warn("Unsuccessful job attempt " + attemptInfo.getAttempt().getId()
-            + " with status " + job.getStatus() + " produced log output as follows: " + attemptInfo.getLogs().getLogLines());
-      }
-    }
-    assertEquals(JobStatus.SUCCEEDED, job.getStatus());
-  }
-
-  private static JobRead waitWhileJobHasStatus(final JobsApi jobsApi, final JobRead originalJob, final Set<JobStatus> jobStatuses)
-      throws InterruptedException, ApiException {
-    return waitWhileJobHasStatus(jobsApi, originalJob, jobStatuses, Duration.ofMinutes(6));
-  }
-
-  @SuppressWarnings("BusyWait")
-  private static JobRead waitWhileJobHasStatus(final JobsApi jobsApi,
-                                               final JobRead originalJob,
-                                               final Set<JobStatus> jobStatuses,
-                                               final Duration maxWaitTime)
-      throws InterruptedException, ApiException {
-    JobRead job = originalJob;
-
-    final Instant waitStart = Instant.now();
-    while (jobStatuses.contains(job.getStatus())) {
-      if (Duration.between(waitStart, Instant.now()).compareTo(maxWaitTime) > 0) {
-        LOGGER.info("Max wait time of {} has been reached. Stopping wait.", maxWaitTime);
-        break;
-      }
-      sleep(1000);
-
-      job = jobsApi.getJobInfo(new JobIdRequestBody().id(job.getId())).getJob();
-      LOGGER.info("waiting: job id: {} config type: {} status: {}", job.getId(), job.getConfigType(), job.getStatus());
-    }
-    return job;
-  }
-
-  public enum Type {
-    SOURCE,
-    DESTINATION
   }
 
 }
